Move responsibility from heat to gridmanager (#335)
